+++ index.browser.js
@@ -1,0 +1,8 @@
+import { Form, Field, List, Submit } from './index.js'
+
+export default {
+	Form,
+	Field,
+	List,
+	Submit
+}