+++ source/plugins/ListPlugin.js
@@ -1,15 +1,54 @@
 import {
 	getListValue,
-	convertListValues
+	convertListValues,
+	removeListAction
 } from './ListPlugin.utility'
 
 export default class ListPlugin {
 	// Tracks `<List/>`s.
 	lists = {}
 
-	initContext(context) {
-		context.onRegisterList = (name, { onReset }) => {
-			this.lists[name] = { onReset }
+	getContextFunctions() {
+		return {
+			onRegisterList: ({ updateState }) => (name, { onReset, state: listState }) => {
+				this.lists[name] = { onReset }
+				updateState((state) => {
+					if (state.listInstanceCounters[name] === undefined) {
+						state.listInstanceCounters[name] = 1
+					} else {
+						state.listInstanceCounters[name]++
+					}
+					state.lists[name] = listState
+				})
+			},
+			onUnregisterList: ({ updateState }) => (name) => {
+				this.lists[name] = undefined
+				// Doesn't reset the list's `state` in form state
+				// because the list might get re-mounted at some other position
+				// in the document during the ongoing React re-render.
+				updateState((state) => {
+					// Doesn't create a new `state` object.
+					// Instead it "mutates" the original one.
+					// The rationale is provided in `actions.js`.
+					state.listInstanceCounters[name]--
+				})
+			},
+			onListStateChange: ({ updateState }) => (name, newState) => {
+				updateState((state) => {
+					// Doesn't create a new `state` object.
+					// Instead it "mutates" the original one.
+					// The rationale is provided in `actions.js`.
+					state.lists[name] = newState
+				})
+			}
+		}
+	}
+
+	getInitialState(state) {
+		return {
+			...state,
+			lists: {},
+			listInstanceCounters: {}
 		}
 	}
 
@@ -25,7 +64,7 @@ export default class ListPlugin {
 
 	onResetField(name, form) {
 		if (this.lists[name]) {
-			const { fields } = form.state
+			const { fields } = form.getState()
 			for (const field of Object.keys(fields)) {
 				if (field.indexOf(`${name}:`) === 0) {
 					form.resetField(field)
@@ -33,6 +72,16 @@ export default class ListPlugin {
 			}
 			this.lists[name].onReset()
 			return true
+		}
+	}
+
+	onUpdate({ getState, dispatch }) {
+		const { listInstanceCounters } = getState()
+		for (const name of Object.keys(listInstanceCounters)) {
+			// Remove unmounted `<List/>`s.
+			if (listInstanceCounters[name] === 0) {
+				dispatch(removeListAction(name))
+			}
 		}
 	}
 }